Feat(ui): Implements work in progress modal
Adds a Work in Progress (WIP) modal to inform users that certain features are under development.
The WipProvider component manages the state of the modal (open/close, message).
The modal is triggered by clicking on specific elements in the Header, Hero, Topics, and Footer components.
Google Analytics events are tracked upon opening and closing the modal.